backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
chromadb==0.4.18
sentence-transformers==2.2.2
datasets==2.15.0
torch==2.1.0
numpy==1.24.3
tqdm==4.66.1
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.0  # For testing
python-jose[cryptography]==3.3.0  # For future JWT support
passlib[bcrypt]==1.7.4  # For future password hashing
tenacity==8.2.3  # For retrying failed operations